restructure files to be installable with setup.py
