+++ .gitignore
@@ -1,3 +1,6 @@
 __pycache__
 *.swp
 *.bak
+/recipemd.egg-info
+/.cache/
+/.idea/
