+++ bin/reprocess-all
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+import sys
+import os
+
+sys.path.insert(1, os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), '..')))
+from recipemd.reprocess_all import main
+main()
