+++ recipemd/errors.py
